Moving structs in lib to model
