+++ src/lib.rs
@@ -14,6 +14,7 @@
 
 pub mod cli;
 pub mod lua;
+mod model;
 
 use cli::{
     bar::{show_msg, MessageLevel, BAR},
@@ -24,50 +25,8 @@ use lua::{
     model::LuaOptions, parsing::files::filename_to_string, run::LuaLoader,
     threads::runner::iter_futures,
 };
-use reqwest::header::HeaderMap;
-
-use std::{
-    path::PathBuf,
-    sync::{Arc, Mutex},
-};
-
-/// Lotus HTTP Options
-#[derive(Clone)]
-pub struct RequestOpts {
-    /// Default Headers
-    pub headers: HeaderMap,
-    /// Custom http Proxy
-    pub proxy: Option<String>,
-    /// Request Timeout
-    pub timeout: u64,
-    /// Limits of http redirects
-    pub redirects: u32,
-}
-
-/// Scanning type For `SCAN_TYPE` in lua scripts
-#[derive(Clone, Copy)]
-pub enum ScanTypes {
-    /// HOSTS Scanning under ID number 1
-    HOSTS,
-    /// URLS Scanning under ID number 2
-    URLS,
-    /// PATHS Scanning under ID number 3
-    PATHS,
-}
-
-pub struct Lotus {
-    /// Script Path
-    pub script_path: PathBuf,
-    /// Output Path
-    pub output: Option<PathBuf>,
-    /// Workers Option
-    pub workers: usize,
-    /// How many url per script
-    pub script_workers: usize,
-    /// Stop After X of errors
-    pub stop_after: Arc<Mutex<i32>>,
-    pub env_vars: serde_json::Value,
-}
+use std::sync::Arc;
+pub use model::{ScanTypes, Lotus, RequestOpts};
 
 impl Lotus {
     /// Run The Lua Script with real target
